[WIP] Lua related functionality has been broken in to own module.
